Small renaming

+++ Scripts/Graphics/GLMesh.cpp
@@ -1,8 +1,6 @@
 #include "GLMesh.h"
-#include "GraphicCore.h"
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
-#include <numeric>
 
 GLMesh::GLMesh() : HaveGPUResources(false) 
 {}
@@ -17,7 +15,7 @@ GLMesh::GLMesh(const std::vector<Vertex3D>& vertexData, const std::vector<GLuint
     glBindBuffer(GL_ARRAY_BUFFER, VBO);
     HaveGPUResources = true;
 
-    IndicesSize = indices.size();
+    IndicesSize = static_cast<int>(indices.size());
     glBufferData(GL_ARRAY_BUFFER, vertexData.size() * sizeof(Vertex3D), &(vertexData[0]), static_cast<GLenum>(drawFlags));
 
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, EBO);
@@ -49,7 +47,7 @@ GLMesh::GLMesh(const std::vector<Vertex2D>& vertexData, const std::vector<GLuint
     glBindBuffer(GL_ARRAY_BUFFER, VBO);
     HaveGPUResources = true;
 
-    IndicesSize = indices.size();
+    IndicesSize = static_cast<int>(indices.size());
     glBufferData(GL_ARRAY_BUFFER, vertexData.size() * sizeof(Vertex2D), &(vertexData[0]), static_cast<GLenum>(drawFlags));
 
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, EBO);
@@ -75,7 +73,7 @@ GLMesh::GLMesh(const std::vector<Vertex>& vertexData, const std::vector<GLuint>&
     glBindBuffer(GL_ARRAY_BUFFER, VBO);
     HaveGPUResources = true;
 
-    IndicesSize = indices.size();
+    IndicesSize = static_cast<int>(indices.size());
     glBufferData(GL_ARRAY_BUFFER, vertexData.size() * sizeof(Vertex), &(vertexData[0]), static_cast<GLenum>(drawFlags));
 
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, EBO);
